Refactor: move error related functions into error module
Change-type: patch
