+++ lib/actions/config.js
@@ -322,7 +322,7 @@ Examples:
 			generateApplicationConfig,
 		} = require('../utils/config');
 		const helpers = require('../utils/helpers');
-		const { exitWithExpectedError } = require('../utils/patterns');
+		const { exitWithExpectedError } = require('../errors');
 
 		if (options.device == null && options.application == null) {
 			exitWithExpectedError(`\
