+++ lib/actions/help.js
@@ -20,7 +20,7 @@ import * as capitano from 'capitano';
 import * as columnify from 'columnify';
 import * as messages from '../utils/messages';
 import { getManualSortCompareFunction } from '../utils/helpers';
-import { exitWithExpectedError } from '../utils/patterns';
+import { exitWithExpectedError } from '../errors';
 import { getOclifHelpLinePairs } from './help_ts';
 
 const parse = object =>
