+++ lib/actions/local/common.js
@@ -1,7 +1,7 @@
 import * as Promise from 'bluebird';
 import * as _ from 'lodash';
 import * as dockerUtils from '../../utils/docker';
-import { exitWithExpectedError } from '../../utils/patterns';
+import { exitWithExpectedError } from '../../errors';
 import { getChalk } from '../../utils/lazy';
 
 export const dockerPort = 2375;
