+++ lib/actions/preload.js
@@ -87,7 +87,7 @@ const getApplicationsWithSuccessfulBuilds = function(deviceType) {
 const selectApplication = function(deviceType) {
 	const visuals = getVisuals();
 	const form = require('resin-cli-form');
-	const { exitWithExpectedError } = require('../utils/patterns');
+	const { exitWithExpectedError } = require('../errors');
 
 	const applicationInfoSpinner = new visuals.Spinner(
 		'Downloading list of applications and releases.',
@@ -116,7 +116,7 @@ const selectApplication = function(deviceType) {
 
 const selectApplicationCommit = function(releases) {
 	const form = require('resin-cli-form');
-	const { exitWithExpectedError } = require('../utils/patterns');
+	const { exitWithExpectedError } = require('../errors');
 
 	if (releases.length === 0) {
 		exitWithExpectedError('This application has no successful releases.');
@@ -263,7 +263,7 @@ Examples:
 		const balenaPreload = require('balena-preload');
 		const visuals = getVisuals();
 		const nodeCleanup = require('node-cleanup');
-		const { exitWithExpectedError } = require('../utils/patterns');
+		const { exitWithExpectedError } = require('../errors');
 
 		const progressBars = {};
