+++ lib/app-capitano.js
@@ -22,7 +22,7 @@ import * as events from './events';
 
 capitano.permission('user', done =>
 	require('./utils/patterns')
-		.exitIfNotLoggedIn()
+		.checkLoggedIn()
 		.then(done, done),
 );
