+++ lib/utils/docker-js.js
@@ -270,7 +270,7 @@ export const createClient = function(opts) {
 };
 
 var ensureDockerSeemsAccessible = function(docker) {
-	const { exitWithExpectedError } = require('./patterns');
+	const { exitWithExpectedError } = require('../errors');
 	return docker
 		.ping()
 		.catch(() =>
